# backend/openapi.yaml
openapi: 3.0.0
info:
  title: 'Chiorino ERP API'
  version: '1'
paths:
...
  /items:
    get:
      tags:
        - Items
        - masterdata_item.php
      summary: 'Item list'
      operationId: 1ce86fb166d24e47d4923801b137ef2a
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  '': { type: object, items: { $ref: '#/components/schemas/ItemResource' } }
                type: array
    post:
      tags:
        - Item
        - masterdata_item_add.php
      summary: 'Create new Item'
      operationId: d4b4bdc9486ae3e0676b9c0e5cc42b61
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - item_desc
                - um
                - item_group
              properties:
                item_desc:
                  type: string
                um:
                  type: string
                item_group:
                  type: string
              type: object
            example:
              item_desc: 'Nome Item'
              um: 'N'
              item_group: NA
              item_subgroup: 0-176
              weight_um: kg
              weight: '10'
              product_line: 0-322
              customs_code: ''
              number_of_plies: '0'
              type: purchased
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResource'
        '403':
          description: Unauthorized
        '500':
          description: 'Internal Error'
...
  /items/autocomplete:
    get:
      tags:
        - Items
      summary: 'Items search for autocomplete'
      operationId: c94a5da900b6fbcd093b934e5bd30105
      parameters:
        -
          name: search
          in: query
          required: true
        -
          name: enabled
          in: query
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  '': { type: object, items: { $ref: '#/components/schemas/ItemResource' } }
                type: array
  '/items/{item}/toggle/{enable}':
    put:
      tags:
        - Item
        - masterdata_item_add.php
      summary: 'Enable ot Disable Item'
      operationId: f56892acd45c13a0002126bd98b50945
      parameters:
        -
          $ref: '#/components/parameters/item'
        -
          $ref: '#/components/parameters/enable'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResource'
        '403':
          description: Unauthorized
        '500':
          description: 'Internal Error'
  '/items/{item}/alternative':
    put:
      tags:
        - Item
        - masterdata_item_alternative.php
      summary: 'Set Item lternative code and description'
      operationId: aabe489439c189bd70138ab1508f386e
      parameters:
        -
          $ref: '#/components/parameters/item'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - altv_code
                - altv_desc
              properties:
                altv_code:
                  type: string
                altv_desc:
                  type: string
              type: object
            example:
              altv_code: 'Alt Code'
              altv_desc: 'Alt Description'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResource'
        '400':
          description: 'Item is not enabled'
        '403':
          description: Unauthorized
        '500':
          description: 'Internal Error'
  /items/groups:
    get:
      tags:
        - Item
        - masterdata_item_add.php
      summary: 'Get available item groups'
      operationId: f0dbf75ba4f332d97bee3a44ef12cd03
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemGroupResource'
        '403':
          description: Unauthorized
        '500':
          description: 'Internal Error'
  '/companies/{company}/items/groups':
    get:
      tags:
        - Item
        - masterdata_item_add.php
      summary: 'Get available item groups for a company'
      operationId: 1f250e29472892c6ea643bb48bb7e84a
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemGroupResource'
        '403':
          description: Unauthorized
        '500':
          description: 'Internal Error'
  '/items/{id}':
    get:
      tags:
        - Item
        - masterdata_item.php
      summary: 'Item show'
      operationId: 07b281b7f44ef98d4983434f2a810753
      parameters:
        -
          $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResource'
        '404':
          description: 'Not Found'
        '403':
          description: Unauthorized
    put:
      tags:
        - Item
        - masterdata_item_add.php
      summary: 'Update Item'
      operationId: af81c8b66985e6d0a51ba49176f19797
      parameters:
        -
          $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - desc
              properties:
                desc:
                  type: string
              type: object
            example:
              item_desc: 'Nome Item'
              um: 'N'
              item_group: NA
              item_subgroup: 0-176
              weight_um: kg
              weight: '10'
              product_line: 0-322
              customs_code: ''
              number_of_plies: '0'
              type: purchased
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResource'
        '403':
          description: Unauthorized
        '500':
          description: 'Internal Error'
    delete:
      tags:
        - Item
        - masterdata_item.php
      summary: 'Delete Item'
      operationId: 788ef26319b67447e0343b342a85ac6d
      parameters:
        -
          $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '403':
          description: Unauthorized
        '404':
          description: 'Not Found'
        '500':
          description: 'Internal Error'
  '/items/{id}/configuration-details':
    get:
      tags:
        - Item
        - masterdata_item.php
      summary: 'Get the product configuration in a human readable way'
      operationId: 50d374d5d454849598fde29509ce9808
      parameters:
        -
          $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResource'
        '404':
          description: 'Not Found'
        '400':
          description: 'Bad request'
        '403':
          description: Unauthorized
  '/items/{id}/stock-limits':
    get:
      tags:
        - Item
        - masterdata_item_stock_limits.php
      summary: 'Get all item stock limits'
      operationId: 86cc22eeb50e3ec416eb3b46d38fefbd
      parameters:
        -
          $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemStockLimitResource'
        '404':
          description: 'Not Found'
        '403':
          description: Unauthorized
    put:
      tags:
        - Item
        - masterdata_item_stock_limits.php
      summary: 'Update Item'
      operationId: 389bbc24c88c92ad98c711c937807567
      parameters:
        -
          $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - IDwarehouse
                - qty_min
                - qty_max
              properties:
                IDwarehouse:
                  type: integer
                qty_min:
                  type: number
                qty_max:
                  type: number
              type: object
            example:
              IDwarehouse: 1
              qty_min: 10
              qty_max: 100
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResource'
        '403':
          description: Unauthorized
        '500':
          description: 'Internal Error'
  '/items/{id}/stock-limits-history':
    get:
      tags:
        - Item
        - masterdata_item_stock_limits.php
      summary: 'Get all item stock limits history'
      operationId: 30e1487b82a1f4c539b1007b9e4bf534
      parameters:
        -
          $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemStockLimitResource'
        '404':
          description: 'Not Found'
        '403':
          description: Unauthorized
...
  /items/types:
    get:
      tags:
        - Item
        - masterdata_item.php
      summary: 'Get list of available item types'
      operationId: a82dfcb538682983803acf5e3c0a8be3
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '403':
          description: Unauthorized
        '404':
          description: 'Not Found'
        '500':
          description: 'Internal Error'
...